# public/openapi.yaml
openapi: 3.0.0
info:
  title: Beantown Baby Diaper Bank API
  version: 1.0.0
  description: OpenAPI docs aligned with implemented Next.js API routes.
servers:
  - url: /api
    description: Internal API
paths:
  /partners:
    get:
      summary: List partners
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: waitlisted
          in: query
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        '200':
          description: Partner list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PartnerListItem'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Create partner
      description: |
        Accepts JSON or multipart/form-data. For multipart, include a `partner`
        JSON string field, optional `logoAction` (keep|replace), and `file` when replacing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartnerPayload'
          multipart/form-data:
            schema:
              type: object
              required: [partner]
              properties:
                partner:
                  type: string
                  description: JSON-serialized CreatePartnerPayload
                logoAction:
                  type: string
                  enum: [keep, replace]
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Partner created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    description: JSON string of the created partner object (BigInt-safe serialization)
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Update partner (body contains id)
      description: |
        Accepts JSON or multipart/form-data. For multipart, include a `partner`
        JSON string field, optional `logoAction` (keep|replace|remove), and `file` when replacing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePartnerPayload'
          multipart/form-data:
            schema:
              type: object
              required: [partner]
              properties:
                partner:
                  type: string
                  description: JSON-serialized UpdatePartnerPayload
                logoAction:
                  type: string
                  enum: [keep, replace, remove]
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Partner updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    description: JSON string of the updated partner object (BigInt-safe serialization)
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    StatusEnum:
      type: string
      enum:
        - active
        - waitlisted
        - inactive
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
    PartnerListItem:
      type: object
      properties:
        id:
          type: integer
        created_at:
          type: string
          format: date-time
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        start_partner:
          type: string
          format: date-time
          nullable: true
        end_partner:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        waitlisted:
          type: boolean
        address:
          type: string
          nullable: true
        coords:
          type: object
          nullable: true
          additionalProperties: true
        logoUrl:
          type: string
          nullable: true
    CreatePartnerPayload:
      type: object
      required: [name, description, status, coordinates, address, cities]
      properties:
        name:
          type: string
        description:
          type: string
        start_partner:
          type: string
          nullable: true
        end_partner:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        coordinates:
          type: object
          additionalProperties: true
        address:
          type: string
        logo:
          type: string
        cities:
          type: array
          items:
            type: object
            required: [city, percentage]
            properties:
              city:
                type: string
              percentage:
                type: number
    UpdatePartnerPayload:
      type: object
      required: [id, name, description, status, coordinates, address]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        start_partner:
          type: string
          nullable: true
        end_partner:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        coordinates:
          type: object
          additionalProperties: true
        address:
          type: string
        logo:
          type: string
...